+++ src/types.ts
@@ -1,5 +1,5 @@
 /**
- * types.ts 11.0.1-dev
+ * types.ts 11.1.2
  * Copyright (c) 2021-2024 Alain Dumesny - see GridStack root license
  */
 
@@ -256,8 +256,9 @@ export interface GridStackOptions {
   rtl?: boolean | 'auto';
 
   /** set to true if all grid items (by default, but item can also override) height should be based on content size instead of WidgetItem.h to avoid v-scrollbars.
-   Note: this is still row based, not pixels, so it will use ceil(getBoundingClientRect().height / getCellHeight()) */
-   sizeToContent?: boolean;
+   * Note: this is still row based, not pixels, so it will use ceil(getBoundingClientRect().height / getCellHeight())
+   */
+  sizeToContent?: boolean;
 
   /**
    * makes grid static (default?: false). If `true` widgets are not movable/resizable.